Update memoryless_excursions.tex
Added a step in this exercise, because it is important to realize that the we cannot just break up this double partial derivative as a product of two partial derivatives (since F_Y(x) depends on x). Added a remark on the notation (which has not been used before, I think,)

Additional suggestion: it might be nice to explicitly check that these densities integrate to 1, possibly creating a new exercise for this.